Update and segment documentation
The documentation is being split-off into targeted sections, keeping the
landing page focused on deploying and configuring the argo-tunnel
controller.

Other changes of note:
 - removes helm chart
 - removes roadmap
 - updated examples: echo
 - adds a series of guides

I understand that removing the help chart is an awkward way to resolve
the request for a daemonset variant. With the updates to the documents,
the installation has been simplified and is not driven through helm.

Once the project is connected to CI the contributing instructions will
be added.

fixes #45
fixes #38
fixes #31