Replace config struct  with option pattern
Remove the config struct and replace it with an option pattern.  This
allows configuration to be both less exposed (doesn't need to be
exported) and partial (only provide the options that are changing,
rather than the entire struct).  The base mechanism will be expanded,
allowing greater control (from either flags or annotations).