allow tunnel to detect and repair in run error
If a tunnel dies during execution, an error will be written to an error
channel.  The change monitors the error channel for issues and triggers
repair as long as the channel is still meant to be running (e.g. is not
stop and did not error during stop).

fix #70
fix #40