replace glog with logrus
Previously, the code base leveraged both loggers, glog and logrus, in
different portions of the code.  THis change consolidates loggers to one
variant, logrus.  The kubernentes client internally uses glog, which
defines the --v flag.  To avoid altering the optiosn for 0.5.x, the
mechanism is preserved.  The value for --v is mapped into a logrus level
and set to the shared logger.